amend! gobject-introspection: 1.76.1 → 1.78.1
gobject-introspection: 1.76.1 → 1.78.1

https://gitlab.gnome.org/GNOME/gobject-introspection/-/compare/1.76.1...1.78.1

NEWS look okay.

* Update the GIR data
* Add GObject as a dependency for the Cairo GIR
* Update regression test suite
* Support different prefix for finding GIR data
* Add GI_GIR_PATH environment variable for controlling GIR paths

The build system changes are suspect – `gir_dir_prefix` is assumed to be relative,
but it takes the value from `datadir` option, which is not generally relative.
Though, currently, the variable is not stringly concatenated anywhere
other than in a code path where its value comes from eponymous option,
which is documented as “Intermediate prefix for gir installation under ${prefix}”,
so it can probably be assumed to be relative there.

Changelog-Reviewed-By: Bobby Rong <rjl931189261@126.com>
Changelog-reviewed-by: Jan Tojnar <jtojnar@gmail.com>